NF: move tests mod to another file
